+++ src/css/Contact.css
@@ -10,7 +10,7 @@
     background-size: cover;
 }
 
-h1{
+.Contact-h1{
     text-align: center;
     color: white;
     font-size: 3rem;
@@ -48,15 +48,9 @@ h1{
    
 } 
 
-h4{
-    font-size: 2rem;
-    font-weight: 500;
-    /*margin-top: 0;*/
-    margin: 0;
-}
 
 /*The paragraph above the contact form*/
-p{
+.Contact-p2{
     color: rgba(255,255,255,.9);
     font-size: 1.2rem;
     font-weight: 300;
@@ -101,7 +95,7 @@ p{
 }
 
 /*No class declared*/
-span{
+.Contact-span{
     font-size: 1.3rem;
     font-weight: 300;
 }
@@ -137,7 +131,7 @@ span{
     padding: 0;
 }
 
-input[type="radio"]:checked + label{
+.Contact-rb/*input[type="radio"]*/:checked + label{
     color: white;
     font-weight: 700;
 }
@@ -160,7 +154,7 @@ button{
     margin-right: 20px;
 }
 
-input[type="text"], input[type="tel"], input[type="email"], textarea{
+.Contact-textinput, .Contact-emailinput, .Contact-telinput,/*input[type="text"], input[type="tel"], input[type="email"],*/ textarea{
     /* background:rgba(255, 255, 255, 0.2); */
     border: none;
     cursor: pointer;
@@ -178,7 +172,7 @@ input[type="text"], input[type="tel"], input[type="email"], textarea{
     border-bottom: 2px solid #05f7ff;
 }
 
-label{
+.Contact-label{
     font-size: .9rem;
     color: rgba(255, 255, 255, .4);
     margin-bottom: 1rem;
